Домашна 4/fundamental_analysis_service/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
selenium==4.10.0
beautifulsoup4==4.12.2
pillow==10.0.1
pytesseract==0.3.10
mammoth==1.4.2
pdf2image==1.16.3
transformers==4.31.0
torch